Refactor: Rename joplinCloudUtils -> joplinOAuthUtils
