+++ packages/app-desktop/app.ts
@@ -48,7 +48,7 @@ import ShareService from '@joplin/lib/services/share/ShareService';
 import checkForUpdates from './checkForUpdates';
 import { AppState } from './app.reducer';
 import syncDebugLog from '@joplin/lib/services/synchronizer/syncDebugLog';
-import { completePendingAuthentication } from '@joplin/lib/services/joplinCloudUtils';
+import { completePendingAuthentication } from '@joplin/lib/services/joplinOAuthUtils';
 import eventManager, { EventName } from '@joplin/lib/eventManager';
 import path = require('path');
 import { afterDefaultPluginsLoaded, loadAndRunDefaultPlugins } from '@joplin/lib/services/plugins/defaultPlugins/defaultPluginsUtils';
